+++ src/Schema/Blueprint.php
@@ -8,5 +8,6 @@ use Illuminate\Database\Schema\Blueprint as BaseBlueprint;
 
 class Blueprint extends BaseBlueprint
 {
+    use BlueprintIndex;
     use BlueprintTypes;
 }