+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- If exists index dropping
+
 ## [0.2.1] - 2021-04-22
 ### Fixed
 - Service Provider no longer extends DatabaseServiceProvider (#1)